Update project structure moving interface to project root
